# apiEscola/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AppEscola API",
        "description": "API REST do sistema multi-tenant AppEscola — gestão de escolas e cursinhos.",
        "contact": {
            "email": "contato@appescola.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:4000",
            "description": "Ambiente local"
        }
    ],
    "paths": {
...
        "/api/guardians": {
            "get": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Listar responsáveis",
                "operationId": "ea638c7c97296ba1fdbca0312d27cce0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de responsáveis"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Criar responsável",
                "operationId": "6571c998951936d28cddd571ddc976b3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Dados do responsável"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Responsável criado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/guardians/{id}": {
            "get": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Exibir responsável",
                "operationId": "e68fa6c0bcf110859a8baea45eabecd9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados do responsável"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Atualizar responsável",
                "operationId": "33485bca0b877db171d765abfe856adf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Campos a atualizar"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Responsável atualizado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Guardians"
                ],
                "summary": "Remover responsável",
                "operationId": "525a7804f7cc034c14701e920d0bcc02",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removido com sucesso"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/school-classes": {
            "get": {
                "tags": [
                    "SchoolClasses"
                ],
                "summary": "Listar turmas",
                "operationId": "03e5143e81a3dbf01822617fea57621b",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "course_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "year",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada de turmas"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SchoolClasses"
                ],
                "summary": "Criar turma",
                "operationId": "778dc8b9836fbb147db1b5c1dd448f63",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Dados da turma"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Turma criada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/school-classes/{id}": {
            "get": {
                "tags": [
                    "SchoolClasses"
                ],
                "summary": "Exibir turma",
                "operationId": "ee8a4fffed93f5fcfd30cb33de4d16f5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dados da turma"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "SchoolClasses"
                ],
                "summary": "Atualizar turma",
                "operationId": "b26afba68e4769316e99d5a1551b0952",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "description": "Campos a atualizar"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Turma atualizada"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SchoolClasses"
                ],
                "summary": "Remover turma",
                "operationId": "0f3435c3e7232e5eb993c7c2a38f414b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Removido com sucesso"
                    },
                    "404": {
                        "description": "Não encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Informe o token retornado pelo endpoint /api/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}